reworking the backend directory structure
