+++ backend-server/utils/json_utils.py
@@ -1,0 +1,2 @@
+events_json_file_path = "../src/data/events.json"
+posts_json_file_path = "../src/data/posts.json"
